requirements.txt
aiohttp==1.0.5
async-timeout==1.1.0
chardet==2.3.0
discord.py==0.15.1
imgurpython==1.1.7
multidict==2.1.4
requests==2.12.3
websockets==3.2
